resolve lazy import race condition (#6272)
some private constants are no longer re-exported to the same spots
